requirements.txt
# Python dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
